More templating.
