+++ cpp/dolfinx/fem/assemble_vector_impl.h
@@ -733,7 +733,7 @@ void apply_lifting(
       for (const std::shared_ptr<const DirichletBC>& bc : bcs1[j])
       {
         bc->mark_dofs(bc_markers1);
-        bc->dof_values(bc_values1);
+        bc->dof_values<T>(bc_values1);
       }
 
       // Modify (apply lifting) vector
